netlink: split rt into per-object classes
route, link and addr become netlink.session subclasses under
lib/netlink/rt/, hung off an rt namespace table: netlink.rt.route():add{}
replaces rt:route_add{} -- the noun moves to the class and the methods
become the verbs, mirroring ip(8) objects, one file per concern. The
rtm_table width probe becomes struct:fieldsize(), shared by the object
classes.
