+++ config.ld
@@ -42,6 +42,9 @@ file = {
 	'./lib/netlink/genl.lua',
 	'./lib/netlink/message.lua',
 	'./lib/netlink/rt.lua',
+	'./lib/netlink/rt/addr.lua',
+	'./lib/netlink/rt/link.lua',
+	'./lib/netlink/rt/route.lua',
 	'./lib/netlink/session.lua',
 	'./lib/luanetfilter.c',
 	'./lib/luanetlink.c',
