+++ lib/netlink/rt.lua
@@ -4,157 +4,36 @@
 --
 
 ---
--- rtnetlink interface for routes, links and addresses. A `netlink.session`
--- specialization over the `NETLINK_ROUTE` protocol: create an instance with
--- `netlink.rt()` and call its methods; the underlying socket is closed by `close()`
--- (or the to-be-closed `__close`). All methods block and require a sleepable
--- runtime.
---
+-- The rtnetlink namespace. Groups the `NETLINK_ROUTE` object classes; each
+-- opens its own request/response session (sleepable).
 -- @module netlink.rt
--- @see netlink.session
---
 
-local session = require("netlink.session")
-local message = require("netlink.message")
-local struct  = require("struct")
+local route = require("netlink.rt.route")
+local link  = require("netlink.rt.link")
+local addr  = require("netlink.rt.addr")
 
-local nl   = require("linux.netlink")
-local rtnl = require("linux.rtnetlink")
-local sk   = require("linux.socket")
-
-local insert = table.insert
-local u32, str = message.u32, message.str
-
-local rtmsg     = struct(rtnl.layout.rtmsg)
-local RTMSG_LEN = rtmsg.size
-local ifinfomsg  = struct(rtnl.layout.ifinfomsg)
-local IFINFO_LEN = ifinfomsg.size
-local ifaddrmsg  = struct(rtnl.layout.ifaddrmsg)
-local IFADDR_LEN = ifaddrmsg.size
-
-local function fieldsize(layout, name)
-	for _, field in ipairs(layout.fields) do
-		if field.name == name then return field.size end
-	end
-end
-
--- ids that do not fit the rtm_table header field go in the RTA_TABLE attribute
-local RTM_TABLE_MAX = (1 << 8 * fieldsize(rtnl.layout.rtmsg, "rtm_table")) - 1
-
-local NEWROUTE, NEWLINK, NEWADDR = rtnl.rtm.NEWROUTE, rtnl.rtm.NEWLINK, rtnl.rtm.NEWADDR
+local rt = {}
 
 ---
--- @type rt
+-- Kernel routing table class.
+-- Open sessions using `netlink.rt.route()`.
+-- @table netlink.rt.route
+-- @see netlink.rt.route
+rt.route = route
 
 ---
--- Creates a new rt object.
--- @function rt:new
--- @tparam[opt] table o an initial object table.
--- @treturn rt the new rt object.
--- @see class
-local rt = session:new{proto = nl.proto.ROUTE}
+-- Network interface (link) class.
+-- Open sessions using `netlink.rt.link()`.
+-- @table netlink.rt.link
+-- @see netlink.rt.link
+rt.link = link
 
 ---
--- Lists all routes from the kernel routing tables.
--- @tparam[opt=AF_UNSPEC] integer family address family.
--- @treturn table list of route tables.
-function rt:route_list(family)
-	local header = rtmsg:pack(family or sk.af.UNSPEC, 0, 0, 0, 0, 0, 0, 0, 0)
-	local routes = {}
-	for _, msg in ipairs(self:dump(rtnl.rtm.GETROUTE, header)) do
-		if msg.type == NEWROUTE then
-			local body = msg.body
-			local fam, dst_len, src_len, tos, tbl, protocol, scope, rtype, flags = rtmsg:unpack(body)
-			local attrs = message.attrs(body, RTMSG_LEN + 1)
-			insert(routes, {
-				family = fam, dst_len = dst_len, src_len = src_len, tos = tos,
-				table = u32(attrs[rtnl.rta.TABLE]) or tbl,
-				protocol = protocol, scope = scope, rtype = rtype, flags = flags,
-				dst = attrs[rtnl.rta.DST], gateway = attrs[rtnl.rta.GATEWAY],
-				oif = u32(attrs[rtnl.rta.OIF]),
-				priority = u32(attrs[rtnl.rta.PRIORITY]),
-			})
-		end
-	end
-	return routes
-end
-
----
--- Lists all network interfaces from the kernel.
--- @treturn table list of link tables.
-function rt:link_list()
-	local header = ifinfomsg:pack(sk.af.UNSPEC, 0, 0, 0, 0)
-	local links = {}
-	for _, msg in ipairs(self:dump(rtnl.rtm.GETLINK, header)) do
-		if msg.type == NEWLINK then
-			local body = msg.body
-			local fam, ltype, ifindex, flags, change = ifinfomsg:unpack(body)
-			local attrs = message.attrs(body, IFINFO_LEN + 1)
-			insert(links, {
-				family = fam, ltype = ltype, ifindex = ifindex,
-				flags = flags, change = change,
-				name = str(attrs[rtnl.ifla.IFNAME]),
-				mtu = u32(attrs[rtnl.ifla.MTU]),
-			})
-		end
-	end
-	return links
-end
-
----
--- Lists all interface addresses from the kernel.
--- @tparam[opt=AF_UNSPEC] integer family address family.
--- @treturn table list of address tables.
-function rt:addr_list(family)
-	local header = ifaddrmsg:pack(family or sk.af.UNSPEC, 0, 0, 0, 0)
-	local addrs = {}
-	for _, msg in ipairs(self:dump(rtnl.rtm.GETADDR, header)) do
-		if msg.type == NEWADDR then
-			local body = msg.body
-			local fam, prefix_len, _, scope, ifindex = ifaddrmsg:unpack(body)
-			local attrs = message.attrs(body, IFADDR_LEN + 1)
-			insert(addrs, {
-				family = fam, prefix_len = prefix_len, scope = scope, ifindex = ifindex,
-				address = attrs[rtnl.ifa.ADDRESS] or attrs[rtnl.ifa.LOCAL],
-				label = str(attrs[rtnl.ifa.LABEL]),
-			})
-		end
-	end
-	return addrs
-end
-
-local function route_attrs(opts)
-	return message.attrs{
-		[rtnl.rta.DST]     = opts.dst,
-		[rtnl.rta.GATEWAY] = opts.gateway,
-		[rtnl.rta.OIF]     = opts.oif,
-		[rtnl.rta.TABLE]   = opts.table and opts.table > RTM_TABLE_MAX and opts.table or nil,
-	}
-end
-
----
--- Adds a route to the kernel routing table.
--- @tparam table opts route parameters: optional `family` (default `AF_INET`),
---   `dst_len`, `dst`, `gateway`, `oif`, `table`, `protocol`, `scope`, `rtype`.
-function rt:route_add(opts)
-	local tbl = opts.table or rtnl.table.MAIN
-	local header = rtmsg:pack(opts.family or sk.af.INET, opts.dst_len or 0, 0, 0,
-		tbl <= RTM_TABLE_MAX and tbl or rtnl.table.UNSPEC, opts.protocol or rtnl.rtprot.STATIC,
-		opts.scope or rtnl.scope.UNIVERSE, opts.rtype or rtnl.rtn.UNICAST, 0)
-	self:talk(NEWROUTE, nl.flag.CREATE | nl.flag.EXCL, header .. route_attrs(opts))
-end
-
----
--- Deletes a route from the kernel routing table.
--- @tparam table opts route parameters: optional `family` (default `AF_INET`),
---   `dst_len`, `dst`, `oif`, `table`.
-function rt:route_del(opts)
-	local tbl = opts.table or rtnl.table.MAIN
-	-- scope NOWHERE is the deletion wildcard: match the route whatever its scope
-	local header = rtmsg:pack(opts.family or sk.af.INET, opts.dst_len or 0, 0, 0,
-		tbl <= RTM_TABLE_MAX and tbl or rtnl.table.UNSPEC, 0, rtnl.scope.NOWHERE, 0, 0)
-	self:talk(rtnl.rtm.DELROUTE, nil, header .. route_attrs(opts))
-end
+-- Interface address class.
+-- Open sessions using `netlink.rt.addr()`.
+-- @table netlink.rt.addr
+-- @see netlink.rt.addr
+rt.addr = addr
 
 return rt
