+++ tests/netlink/route_adddel.lua
@@ -13,29 +13,29 @@ local DST     = string.char(192, 0, 2, 0) -- 192.0.2.0 (TEST-NET-1), network byt
 local DST_LEN = 24
 local LO      = 1                          -- loopback ifindex
 
-local rt <close> = netlink.rt()
+local route <close> = netlink.rt.route()
 
 local function present()
-	for _, route in ipairs(rt:route_list(af.INET)) do
-		if route.table == TABLE and route.dst == DST and route.dst_len == DST_LEN then
+	for _, entry in ipairs(route:list(af.INET)) do
+		if entry.table == TABLE and entry.dst == DST and entry.dst_len == DST_LEN then
 			return true
 		end
 	end
 	return false
 end
 
-rt:route_add{family = af.INET, dst = DST, dst_len = DST_LEN, oif = LO,
+route:add{family = af.INET, dst = DST, dst_len = DST_LEN, oif = LO,
 	table = TABLE, scope = scope.LINK}
 assert(present(), "route not found after add")
 print("netlink route_adddel: added")
 
 -- route_add sends NLM_F_EXCL, so adding the same route again must raise
 -- (EEXIST), surfaced by check_error from the kernel acknowledgment
-assert(not pcall(rt.route_add, rt, {family = af.INET, dst = DST, dst_len = DST_LEN,
+assert(not pcall(route.add, route, {family = af.INET, dst = DST, dst_len = DST_LEN,
 	oif = LO, table = TABLE, scope = scope.LINK}), "duplicate add should raise")
 print("netlink route_adddel: duplicate add raises")
 
-rt:route_del{family = af.INET, dst = DST, dst_len = DST_LEN, oif = LO, table = TABLE}
+route:del{family = af.INET, dst = DST, dst_len = DST_LEN, oif = LO, table = TABLE}
 assert(not present(), "route still present after del")
 print("netlink route_adddel: deleted")
